slamd-96: Renamed discovery_experiment.py -> discovery_conductor.py
